requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework
markdown
channels
daphne
channels_redis
djangorestframework-simplejwt
gunicorn==23.0.0
Markdown==3.10
packaging==25.0
psycopg2-binary==2.9.11
sqlparse==0.5.4
